stackdio/stackdio/requirements/base.txt
Django==1.5.1
Jinja2==2.6
M2Crypto==0.21.1
MySQL-python==1.2.4
PyYAML==3.10
South==0.7.6
amqp==1.0.11
anyjson==0.3.3
billiard==2.7.3.28
celery==3.0.19
distribute==0.6.31
django-celery==3.0.17
django-filter==0.6
djangorestframework==2.3.1
ipython==0.13.2
kombu==2.5.10
msgpack-python==0.3.0
pycrypto==2.6
python-dateutil==2.1
pytz==2013b
pyzmq==13.1.0
readline==6.2.4.1
requests==1.2.0
salt==0.15.1
salt-cloud==0.8.8
see==1.0.1
simplejson==3.2.0
six==1.3.0
django-braces==1.0.0
django-extensions=1.1.1
